Changed server implementation from express to yoga and schema definition from Graphql JS to Graphql DSL
also, changed the schema definition
