+++ server/env.example
@@ -1,1 +1,3 @@
-MONGO_URI = 'mongodb+srv://<Username>:<passwords>@nodeexpressprojects.4qfr9vo.mongodb.net/TopCrew?retryWrites=true&w=majority'
+MONGO_URI = 
+JWT_SECRET = 
+PORT =
