+++ server/package.json
@@ -11,6 +11,7 @@
     "graphql": "^16.6.0",
     "graphql-middleware": "^6.1.33",
     "graphql-shield": "^7.6.5",
+    "graphql-yoga": "^3.7.0",
     "jsonwebtoken": "^9.0.0",
     "mongoose": "^6.9.1"
   },
