+++ server/src/index.js
@@ -1,23 +1,17 @@
-const express = require('express');
-const { graphqlHTTP } = require('express-graphql');
-const mongoose = require('mongoose')
-const schema = require("./schema/schema")
-const connectDB = require('./config/db')
+const { createServer } = require("node:http") // server creater for node js
+const { createYoga } = require("graphql-yoga")
+const { schema } = require("./schema/schema")
+const connectDB = require("./config/db")
 require('dotenv').config()
 
-const app = express();
+const yoga = createYoga({ schema });
+const server = createServer(yoga)
 
-// Middleware to implement GraphQL
-app.use("/gql", graphqlHTTP({
-  schema: schema,
-  graphiql: true
-}))
-
-const startApp = async()=>{
+const startApp = async () => {
+  // Connect to the database
   await connectDB();
-
   // Start the server
-  app.listen(process.env.PORT, ()=>{
+  server.listen(process.env.PORT, () => {
     console.log(`Server listening on port ${process.env.PORT}`)
   })
 }
