--- server/src/resolvers/resolvers.js
@@ -1,7 +1,0 @@
-// const resolvers = {
-//   hello : ()=>{
-//     return "Hello World"
-//   }
-// }
-
-// module.exports = resolvers;
