+++ server/src/schema/resolvers/resolvers.js
@@ -1,0 +1,17 @@
+const { newUserSignUp, userLogin } = require("../resolvers/user")
+const { getAllUsersProfile, getUserProfile } = require("../resolvers/profile")
+
+const resolvers = {
+  Query: {
+    getAllUsersProfile: getAllUsersProfile,
+    getUserProfile: getUserProfile
+  },
+  Mutation: {
+    newUserSignUp: newUserSignUp,
+    userLogin: userLogin
+  }
+}
+
+module.exports = {
+  resolvers
+}
