--- server/src/schema/types/types.js
@@ -1,20 +1,0 @@
-const {
-  GraphQLObjectType,
-  GraphQLID,
-  GraphQLString,
-} = require('graphql');
-
-const UserType = new GraphQLObjectType({
-  name: "User",
-  fields: ()=>({
-    id: { type: GraphQLID},
-    name: { type: GraphQLString},
-    phone: { type: GraphQLString},
-    email: { type: GraphQLString},
-    prodileId: {type: GraphQLString}
-  })
-})
-
-module.exports = {
-  UserType
-}
